packages/backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# OAuth
httpx==0.26.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
msal==1.26.0

# Task Queue
celery[redis]==5.3.5
redis==5.0.1

# AWS
boto3==1.34.19

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
types-redis==4.6.0.11
types-passlib==1.7.7.13